Added Mint and Used total values
Changed the the functionality of adding stamps to collection.  Before each user was responsible for adding the value of the stamp that they're adding.
Now the admin of the site is responsible for adding the stamp's face, mint and used values.
When a user adds a stamp to their collection, they select the grading of the stamp, and then their collection is calculated from the relevant value (mint or used)